:white_check_mark:[#1707] move tests
