+++ src/openzaak/components/documenten/tests/admin/test_eio.py
@@ -2,6 +2,7 @@
 # Copyright (C) 2020 Dimpact
 from django.test import tag
 from django.urls import reverse
+from django.utils.translation import gettext_lazy as _
 
 from django_webtest import WebTest
 from maykin_2fa.test import disable_admin_mfa
@@ -9,7 +10,10 @@ from webtest import Upload
 
 from openzaak.accounts.tests.factories import SuperUserFactory
 from openzaak.components.catalogi.tests.factories import InformatieObjectTypeFactory
-from openzaak.components.documenten.models import EnkelvoudigInformatieObject
+from openzaak.components.documenten.models import (
+    EnkelvoudigInformatieObject,
+    EnkelvoudigInformatieObjectCanonical,
+)
 
 from ..factories import EnkelvoudigInformatieObjectCanonicalFactory
 
@@ -108,3 +112,53 @@ class EnkelvoudigInformatieObjectAdminTests(WebTest):
 
         self.assertEqual(response.status_code, 200)
         self.assertIn("Je moet een informatieobjecttype opgeven", response.text)
+
+
+@disable_admin_mfa()
+class EnkelvoudigInformatieObjectCanonicalAdminTests(WebTest):
+    @classmethod
+    def setUpTestData(cls):
+        cls.user = SuperUserFactory.create()
+
+    def setUp(self):
+        super().setUp()
+
+        self.app.set_user(self.user)
+
+    def test_eio_add_no_version(self):
+
+        self.assertEqual(EnkelvoudigInformatieObjectCanonical.objects.count(), 0)
+
+        add_url = reverse("admin:documenten_enkelvoudiginformatieobjectcanonical_add")
+        get_response = self.app.get(add_url)
+
+        response = get_response.form.submit()
+        self.assertEqual(response.status_code, 200)
+
+        version_form = response.context["inline_admin_formsets"][0].forms[0]
+
+        self.assertEqual(
+            version_form.errors["bronorganisatie"], [_("This field is required.")]
+        )
+        self.assertEqual(
+            version_form.errors["creatiedatum"], [_("This field is required.")]
+        )
+        self.assertEqual(version_form.errors["titel"], [_("This field is required.")])
+        self.assertEqual(version_form.errors["auteur"], [_("This field is required.")])
+        self.assertEqual(version_form.errors["taal"], [_("This field is required.")])
+        self.assertEqual(
+            version_form.errors["bestandsomvang"], [_("This field is required.")]
+        )
+        self.assertEqual(version_form.errors["inhoud"], [_("This field is required.")])
+        self.assertEqual(
+            version_form.errors["__all__"],
+            [
+                _("Constraint “%(name)s” is violated.")
+                % {
+                    "name": "documenten_enkelvoudiginformatieobject__informatieobjecttype_or"
+                    "__informatieobjecttype_base_url_filled"
+                }
+            ],
+        )
+        # should still be zero
+        self.assertEqual(EnkelvoudigInformatieObjectCanonical.objects.count(), 0)
